Moved catalog restapi to search-api, remved references to search-old
