+++ applications/harvester-api/src/main/resources/application.yml
@@ -45,7 +45,7 @@ application:
 
 ---
 spring:
-  profiles: develop, unit-integration
+  profiles: develop, docker, unit-integration
   mail:
     host: smtp.gmail.com
     port: 587
