+++ applications/nginx-search/src/main/docker/nginx.conf
@@ -110,11 +110,7 @@ http {
         }
 
         location /catalogs {
-             proxy_pass http://search-old:8080;
-        }
-
-        location /theme {
-            proxy_pass http://search-old:8080;
+             proxy_pass http://search-api:8080;
         }
 
         location /publisher {
@@ -126,13 +122,7 @@ http {
          }
 
         location /harvest/catalog {
-                     proxy_pass http://search-api:8080;
-        }
-
-        location /css {
-
-            #location ~* \.(js|png|jpg|ttf|svg|eot|woff|woff2|css)$ {
-            proxy_pass http://search-old:8080;
+            proxy_pass http://search-api:8080;
         }
 
         location /api-docs/reference-data {
