+++ applications/pom.xml
@@ -23,7 +23,6 @@
         <module>reference-data</module>
 
         <module>nginx-search</module>
-        <module>search-old</module>
         <module>search-api</module>
         <module>search</module>
