+++ applications/search-api/src/main/java/no/dcat/portal/query/SupportedFormat.java
@@ -1,4 +1,4 @@
-package no.dcat.portal.webapp;
+package no.dcat.portal.query;
 
 import org.apache.jena.riot.Lang;
