+++ applications/search-api/src/main/resources/sparql/allcatalogs.sparql
