+++ applications/search-api/src/main/resources/sparql/dataset.sparql
