+++ applications/search-api/src/test/java/no/dcat/portal/query/CatalogServiceTest.java
@@ -1,4 +1,4 @@
-package no.dcat.portal.webapp;
+package no.dcat.portal.query;
 
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecution;
@@ -7,8 +7,6 @@ import org.apache.jena.query.QueryFactory;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.riot.RDFDataMgr;
-import org.apache.jena.sparql.algebra.optimize.TransformOrderByDistinctApplication;
-import org.apache.jena.sparql.engine.QueryEngineFactory;
 import org.apache.jena.sparql.engine.http.QueryExceptionHTTP;
 import org.junit.Before;
 import org.junit.Test;
@@ -40,20 +38,20 @@ import static org.mockito.Mockito.spy;
 /**
  * Created by dask on 20.03.2017.
  */
-public class PortalRestControllerTest {
-    private static Logger logger = LoggerFactory.getLogger(PortalRestControllerTest.class);
+public class CatalogServiceTest {
+    private static Logger logger = LoggerFactory.getLogger(CatalogServiceTest.class);
 
-    private PortalRestController portal;
+    private CatalogService portal;
 
     @Before
     public void setup() {
-        portal = new PortalRestController();
+        portal = new CatalogService();
     }
 
 
     @Test
     public void getCatalogsOK() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
 
         // Load testdatasett in model
         Resource mResource = new ClassPathResource("data.ttl");
@@ -75,7 +73,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogsFailsOnIOError() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
 
         doThrow(new IOException("force exception")).when(spy).read(anyObject());
 
@@ -87,7 +85,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogsFailsOnNoCatalogsFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
 
         // read modelfile with no catalog in
         Resource mResource = new ClassPathResource("data-no-catalog.ttl");
@@ -110,7 +108,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogDcatOk() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn("DCAT format").when(spy).findResourceById(anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -121,7 +119,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getDatasetDcatFormatsOK() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn("DCAT format").when(spy).findResourceById(anyString(), anyString(), anyString());
 
         final String[] acHeaders = {"application/ld+json", "json", "ld+json", "application/rdf+xml", "rdf", "text/turtle", "turtle"};
@@ -141,7 +139,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogDcatWrongAcceptHeader() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn("DCAT format").when(spy).findResourceById(anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -152,7 +150,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogDcatWrongFormat() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn("DCAT format").when(spy).findResourceById(anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -164,7 +162,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogDcatIdNotFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn(null).when(spy).findResourceById(anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -175,7 +173,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogDcatIdThrowsNotFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doThrow(new NoSuchElementException("NSEE")).when(spy).findResourceById(anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -186,7 +184,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void findResourceByIdOK() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         Resource mResource = new ClassPathResource("data.ttl");
         org.apache.jena.query.Dataset dataset = RDFDataMgr.loadDataset(mResource.getURL().toString());
         Model model = ModelFactory.createUnion(ModelFactory.createDefaultModel(), dataset.getDefaultModel());
@@ -210,7 +208,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void findResourceWithunknownIdReturnsNull() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         Resource mResource = new ClassPathResource("data.ttl");
         org.apache.jena.query.Dataset dataset = RDFDataMgr.loadDataset(mResource.getURL().toString());
         Model model = ModelFactory.createUnion(ModelFactory.createDefaultModel(), dataset.getDefaultModel());
@@ -232,7 +230,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void exportDatasetFusekiNotFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn(null).when(spy).invokeFusekiQuery(anyString(), anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getDatasetDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -243,7 +241,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void exportCatalogFusekiNotFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doReturn(null).when(spy).invokeFusekiQuery(anyString(), anyString(), anyString(), anyString());
 
         ResponseEntity<String> response = spy.getCatalogDcat("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -254,7 +252,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void invokeFusekiQueryThrowsExceptionSinceNoSparqlQueryFileIsFound() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
 
         ResponseEntity<String> response = spy.invokeFusekiQuery("http://data.brreg.no/datakatalog/katalog/974761076/5",
                 null, "application/ld+json", "sparql/nofile.sparql");
@@ -265,7 +263,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void invokeFusekiQueryThrowsExceptionbecauseOfIOErrorInFuseki() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doThrow(new QueryExceptionHTTP(404, "force exception")).when(spy).getQueryExecution(anyObject());
 
         ResponseEntity<String> response = spy.invokeFusekiQuery("http://data.brreg.no/datakatalog/katalog/974761076/5",
@@ -276,7 +274,7 @@ public class PortalRestControllerTest {
 
     @Test
     public void getCatalogsThrowsExceptionBecauseOfIOErrorInFuseki() throws Throwable {
-        PortalRestController spy = spy(portal);
+        CatalogService spy = spy(portal);
         doThrow(new QueryExceptionHTTP(406, "force exception")).when(spy).getQueryExecution(anyObject());
 
         ResponseEntity<String> response = spy.getCatalogs("*/*");
