+++ docker-compose.override.yml
@@ -23,18 +23,6 @@ services:
      - application.elasticsearchHost=elasticsearch
      - application.elasticsearchPort=9300
 
-  search-old:
-    restart: always
-    ports:
-      - 127.0.0.1:3001:8080
-    environment:
-     - JAVA_OPTS= -ea -Djava.security.egd=file:/dev/./urandom -Xmx256M -Dspring.profiles.active=docker
-     - application.elasticsearchHost=elasticsearch
-     - application.elasticsearchPort=9300
-     - application.queryService=http://search-api:8080
-     - application.queryServiceExternal=http://localhost:8083
-     - application.fusekiService=http://fuseki:8080/fuseki
-
   nginx-search:
     image: dcatno/nginx-search:latest
     restart: always
