+++ docker-compose.yml
@@ -10,15 +10,6 @@ services:
     networks:
       platform:
 
-  search-old:
-    image: dcatno/search-old:latest
-    depends_on:
-      - search-api
-      - fuseki
-      - reference-data
-    networks:
-      platform:
-
   search-api:
     image: dcatno/search-api:latest
     depends_on:
@@ -31,7 +22,6 @@ services:
     depends_on:
       - search
       - search-api
-      - search-old
     networks:
       platform:
 
@@ -62,7 +52,6 @@ services:
       logging:
     depends_on:
      - elasticsearch
-    depends_on:
      - fuseki
      - harvester-api
