+++ runCreateAllServicesInOpenshift.sh
@@ -53,7 +53,7 @@ deploymentmode=$4
 
 
 #midlertidig tatt ut reference-data pga at mounting av volum ikke virker
-services="elasticsearch fuseki registration registration-react registration-auth registration-api registration-validator gdoc harvester harvester-api search search-old search-api nginx nginx-registration nginx-search"
+services="elasticsearch fuseki registration registration-react registration-auth registration-api registration-validator gdoc harvester harvester-api search search-api nginx nginx-registration nginx-search"
 
 for i in $services
 do
