+++ runCreateServiceInOpenshift.sh
@@ -327,13 +327,15 @@ then
     if [ $deploymode = recreateServices ]
     then
         profile=prod
-        createOpenshiftService search-old
-        oc env dc/search-old search_referenceDataExternalUrl=https://reference-data-fellesdatakatalog-$environment.$cluster.brreg.no search_queryServiceExternal=https://search-api-fellesdatakatalog-$environment.$cluster.brreg.no
-        exposeService search-old
-        oc expose dc/search-old --port=8080
+        #delete search-old
+        oc delete all -l search-old
+        #createOpenshiftService search-old
+        ##oc env dc/search-old search_referenceDataExternalUrl=https://reference-data-fellesdatakatalog-$environment.$cluster.brreg.no search_queryServiceExternal=https://search-api-fellesdatakatalog-$environment.$cluster.brreg.no
+        #exposeService search-old
+        #oc expose dc/search-old --port=8080
     else
         # deploymentmode = onlyDeployImages
-        deployNewDockerImage search-old
+        #deployNewDockerImage search-old
     fi
 
 elif [ $service = search ]
